// src/app/core/interfaces/iuser.ts
export interface IUser {
  uuid?: string;
  first_name?: string;
  last_name?: string;
  created_on?: Date;
  email?: string;
  password?: string;
  token?: string;
  avatar?: string;
  banner?: string;
  biography?: string;
  location?: string;
  occupation?: string;
  website?: string;
  image?: any;
}